_notes/Public/Talks/MongoDB/real-time-inference-change-streams/src/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
boto3-stubs==1.37.0
botocore-stubs==1.36.26
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
filelock==3.17.0
flask==3.1.0
flask-pymongo==3.0.1
flask-socketio==5.5.1
frozenlist==1.5.0
fsspec==2025.2.0
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
levenshtein==0.26.1
litellm==1.61.15
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
mypy-boto3-bedrock-runtime==1.37.0
openai==1.64.0
opik==1.5.1
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.0
pygments==2.19.1
pymongo==4.11.1
pytest==8.3.4
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pyyaml==6.0.2
rapidfuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
sentry-sdk==2.22.0
setuptools==75.8.0
simple-websocket==1.1.0
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
types-awscrt==0.23.10
types-s3transfer==0.11.2
typing-extensions==4.12.2
urllib3==2.3.0
uuid6==2024.7.10
werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope-event==5.0
zope-interface==7.2
